* LongoMatch/LongoMatch.mdp: * CesarPlayer/Gui/PlayerBin.cs: Cast int to double for division * LongoMatch/DB/TeamTemplate.cs: Add function to set the template dinamically * LongoMatch/gtk-gui/gui.stetic: * CesarPlayer/gtk-gui/gui.stetic: * LongoMatch/Gui/PlayerProperties.cs: We can set images!!!! * LongoMatch/Gui/TeamTemplateEditor.cs: New team template editor * LongoMatch/Gui/TeamTemplateWidget.cs: * LongoMatch/Gui/FileDescriptionWidget.cs: * LongoMatch/gtk-gui/LongoMatch.Gui.Dialog.TeamTemplateEditor.cs: * LongoMatch/gtk-gui/LongoMatch.Gui.Component.PlayerProperties.cs: * LongoMatch/gtk-gui/LongoMatch.Gui.Component.TeamTemplateWidget.cs: * LongoMatch/gtk-gui/LongoMatch.Gui.Component.FileDescriptionWidget.cs:

+++ CesarPlayer/Gui/PlayerBin.cs
@@ -112,7 +112,7 @@ namespace LongoMatch.Gui
 				if (pixbuf != null){
 					h = pixbuf.Height;
 					w = pixbuf.Width;
-					rate = w/h;
+					rate = (double)w/(double)h;
 					return pixbuf.ScaleSimple(THUMBNAIL_WIDTH,(int)(THUMBNAIL_WIDTH/rate),InterpType.Bilinear);
 				}
 				else return null;
